refactor: move sliding window rate counter with min data point limit type to independent file
